backend/yandex-auth/requirements.txt
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
requests>=2.31.0
